+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- All relative hrefs are made absolute on item download
+
 ## [0.0.2] - 2023-05-25
 
 ### Added
@@ -15,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Fixed
 
-- Cleaning up connections is easier thanks to `__aexit__` implementations.
+- Cleaning up connections is easier thanks to `__aexit__` implementations
 
 ## [0.0.1] - 2023-05-24